+++ Entities/HDFirearm.h
@@ -873,13 +873,14 @@ protected:
     // Range of spread in ang vel of ejected shells, in one direction
     float m_ShellAngVelRange;
 	float m_ShellVelVariation; //!< The velocity variation scalar of ejected shells.
+	// The amount of screenshake that recoil causes
+	float m_RecoilScreenShakeAmount;
     // The muzzle velocity the AI use when aiming this weapon
     float m_AIFireVel;
     // The bullet life time the AI use when aiming this weapon
     unsigned long m_AIBulletLifeTime;
     // The bullet acc scalar the AI use when aiming this weapon
     float m_AIBulletAccScalar;
-
     
     // Whether at least one round has already been
     // fired during the current activation.