+++ Entities/MOSRotating.h
@@ -676,6 +676,12 @@ ClassInfoGetters;
     /// <param name="newGibBlastStrength">The new gib blast strength to use.</param>
     void SetGibBlastStrength(float newGibBlastStrength) { m_GibBlastStrength = newGibBlastStrength; }
 
+    /// <summary>
+	/// Gets the amount of screenshake this will cause upon gibbing.
+	/// </summary>
+	/// <returns>The amount of screenshake this will cause when gibbing. If -1, this is calculated automatically.</returns>
+	float GetGibScreenShakeAmount() const { return m_GibScreenShakeAmount; }
+
 	/// <summary>
 	/// Gets a const reference to the list of Attachables on this MOSRotating.
 	/// </summary>
@@ -997,6 +1003,7 @@ protected:
     float m_GibImpulseLimit;
 	int m_GibWoundLimit; //!< The number of wounds that will gib this MOSRotating. 0 means that it can't be gibbed via wounds.
     float m_GibBlastStrength; //!< The strength with which Gibs and Attachables will get launched when this MOSRotating is gibbed.
+    float m_GibScreenShakeAmount; //!< Determines how much screenshake this causes upon gibbing.
 	float m_WoundCountAffectsImpulseLimitRatio; //!< The rate at which this MOSRotating's wound count will diminish the impulse limit.
 	bool m_DetachAttachablesBeforeGibbingFromWounds; //!< Whether to detach any Attachables of this MOSRotating when it should gib from hitting its wound limit, instead of gibbing the MOSRotating itself.
 	bool m_GibAtEndOfLifetime; //!< Whether or not this MOSRotating should gib when it reaches the end of its lifetime, instead of just deleting.