+++ CHANGELOG.md
@@ -468,6 +468,26 @@ This can be accessed via the new Lua (R/W) `SettingsMan` property `AIUpdateInter
 - Added `ACrab` INI properties for setting individual foot `AtomGroup`s, as opposed to setting the same foot `AtomGroup`s for both `Legs` on the left or right side.  
 	These are `LeftFGFootGroup`, `LeftBGFootGroup`, `RightFGFootGroup` and `RightBGFootGroup`.
 
+- Added screen-shake. The screen-shake strength can be tweaked or disabled in the options menu.  
+	New `MOSRotating` INI property `GibScreenShakeAmount`, which determines how much this will shake the screen when gibbed. This defaults to automatically calculating a screen-shake amount based on the energy involved in the gib.  
+	New `HDFirearm` INI property `RecoilScreenShakeAmount`, which determines how much this weapon whill shake the screen when fired. This defaults to automatically calculating a screen-shake amount based on the recoil energy.  
+
+	New `Settings.ini` screen-shake properties:  
+	`ScreenShakeStrength` - a global multiplier applied to screen shaking strength.  
+	`ScreenShakeDecay` - how quickly screen shake falls off.  
+	`MaxScreenShakeTime` - the amount of screen shake time, i.e. the maximum number of seconds screen shake will happen until ScreenShakeDecay reduces it to zero.  
+	`DefaultShakePerUnitOfGibEnergy` - how much the screen should shake per unit of energy from gibbing (i.e explosions), when the screen shake amount is auto-calculated.  
+	`DefaultShakePerUnitOfRecoilEnergy` - how much the screen should shake per unit of energy for recoil, when the screen shake amount is auto-calculated.  
+	`DefaultShakeFromRecoilMaximum` - the maximum amount of screen shake recoil can cause, when the screen shake amount is auto-calculated. This is ignored by per-firearm shake settings.
+
+- Added new Lua manager `CameraMan`, to handle camera movement.
+	New Lua functions on CameraMan:
+	```lua
+	AddScreenShake(screenShakeAmount, screen); -- Can be used to shake a particular screen.
+	AddScreenShake(screenShakeAmount, position); -- Applies screenshake at a position in the game world. All screens looking near this position will have their screen shaken.
+	```
+	Several `SceneMan` Lua functions have been moved into CameraMan. For the full list, see the Changed section below.
+
 - Added alternative `Actor` Lua function `RemoveInventoryItem(moduleName, presetName)`, that lets you specify the module and preset name of the inventory item, instead of just the preset name.
 
 - Added alternative `AHuman` Lua function `EquipNamedDevice(moduleName, presetName, doEquip)`, that lets you specify the module and preset name of the `HeldDevice` to equip, instead of just the preset name.
@@ -616,6 +636,18 @@ This can be accessed via the new Lua (R/W) `SettingsMan` property `AIUpdateInter
 - `BitmapPrimitive` drawing functions now accept `MOSprite` instead of `Entity` for the object they get the bitmap to draw from.  
 	This changes nothing regarding the bindings, but will now print an error to the console when attempting to draw a non-`MOSprite` based object (e.g. `MOPixel`), instead of silently skipping it.
 
+- The following `SceneMan` functions have been moved to `CameraMan`:
+	```lua
+	SetOffset(offsetVector, screenId);
+	GetScreenOcclusion(screenId);
+	SetScreenOcclusion(occlusionVector, screenId);
+	GetScrollTarget(screenId);
+	SetScrollTarget(targetPosition, screenId);
+	TargetDistanceScalar(point);
+	CheckOffset(screenId);
+	SetScroll(center, screenId);
+	```
+
 </details>
 
 <details><summary><b>Fixed</b></summary>
@@ -649,6 +681,8 @@ This can be accessed via the new Lua (R/W) `SettingsMan` property `AIUpdateInter
 - Removed `Activity` Lua function `EnteredOrbit`. This tells the `Activity` to consider an `ACraft` as having entered orbit, and should never actually have been accessible to Lua.
 
 - Removed `OnPieMenu` listeners for `Activity`s and `GlobalScript`s, and removed the `ProvidesPieMenuContext` concept and everything around it. These things should no longer be necessary since you can modify `PieMenu`s on the fly at any time, and they made this already complex set of code even more complicated.
+
+- Removed `SceneMan` Lua functions `SetOffsetX(x, screenId)` and `SetOffsetY(y, screenId)`. Use `CameraMan:SetOffset(offsetVector, screenId)` instead.
 
 </details>